[feature/fleet-sync] refactor: fmt.rs, and the copies it retires
The first of the pure leaves out of lib.rs, and the reason to do it first is
what was hiding in the duplication:

- Three now_ms. One returned u64 and two returned i64, so every consumer of
  the unsigned one cast at the call site to subtract two of them. It is
  signed now and eleven casts went with it.
- Four "the first of this month", three of them a format("%Y-%m-01") round
  trip through a string that fell back to *today* if the parse failed - a bug
  in which would have read as "the month starts today". with_day(1) cannot
  fail.
- Two duration formatters with the same arithmetic side by side. One helper
  taking the number of components, so the reset time keeps its three and the
  burn projection its two: a projection is a guess and does not earn a
  minutes field, the reset is the provider's own number and does.

Every string the two formatters produced is unchanged; the tests that pin
them did not move.

+++ crates/tokengauge-core/src/panel.rs
@@ -233,7 +233,7 @@ pub fn panel_spec(row: &ProviderRow) -> Vec<Section> {
 /// mixed per-provider setup readable without inventing a marker for it.
 fn device_rows(cost: &CostInfo) -> Vec<PanelRow> {
     let max = cost.by_device.first().map(|d| d.tokens).unwrap_or(0);
-    let now_ms = crate::now_ms() as i64;
+    let now_ms = crate::now_ms();
     cost.by_device
         .iter()
         .map(|device| {
@@ -688,7 +688,7 @@ mod tests {
                     label: "desktop".into(),
                     tokens: 900_000_000,
                     usd: 700.0,
-                    updated_at_ms: crate::now_ms() as i64,
+                    updated_at_ms: crate::now_ms(),
                     partial: false,
                     is_local: true,
                 },
@@ -697,7 +697,7 @@ mod tests {
                     label: "laptop".into(),
                     tokens: 500_000_000,
                     usd: 350.91,
-                    updated_at_ms: crate::now_ms() as i64 - 7_200_000,
+                    updated_at_ms: crate::now_ms() - 7_200_000,
                     partial: true,
                     is_local: false,
                 },
